All tests work now

+++ FileServerApp/file_service.py
@@ -1,8 +1,7 @@
 import logging
 import os
-from collections import OrderedDict
 
-from FileServerApp.config import DEFAULT_FILE_CONTENT, LOG_LEVEL, LOG_FORMAT
+from FileServerApp.config import DEFAULT_FILE_CONTENT, LOG_LEVEL, LOG_FORMAT, ENVVAR_NAME_ROOT, FILE_EXTENSION
 from FileServerApp.crypto import AESCipher
 from FileServerApp.utils import generate_random_file_name, check_file_existence, convert_datetime
 
@@ -23,34 +22,32 @@ ch.setFormatter(formatter)
 logger.addHandler(ch)
 
 
-def read_file(name):
+def read_file(file_name):
     """Return content from the given file
 
-    :param name: string with <file name + file extension>
+    :param file_name: string with <file name>
     :return: string with <file content>
     """
-    file_path = check_file_existence(name)
+    file_path = os.path.join(os.getenv(ENVVAR_NAME_ROOT), file_name + FILE_EXTENSION)
+    check_file_existence(file_path)
     aes = AESCipher()
 
     with open(file_path, "rb") as r_file:
-        content = aes.decrypt(r_file, "AES_{}".format(name))
+        content = aes.decrypt(r_file, "AES_{}".format(file_name))
 
     return content.decode("utf-8")
 
 
-def delete_file(name):
+def delete_file(file_name):
     """Remove file if exists
 
-    :param name: string with <file name + file extension>
+    :param file_name: string with <file name + file extension>
     :return: None
     """
-    try:
-        file_path = check_file_existence(name)
-    except NameError:
-        return
-
-    os.remove(file_path)
-    logger.info("File {} was removed".format(name))
+    file_path = os.path.join(os.getenv(ENVVAR_NAME_ROOT), file_name + FILE_EXTENSION)
+    if os.path.isfile(file_path):
+        os.remove(file_path)
+        logger.info("File {} was removed".format(file_name))
 
 
 def create_file():
@@ -59,10 +56,10 @@ def create_file():
     :return: filename and key_filename, both with extension
     """
     file_name = generate_random_file_name()
-    dst_path = merge_filename_with_root(file_name)
+    file_path = os.path.join(os.getenv(ENVVAR_NAME_ROOT), file_name + FILE_EXTENSION)
     aes = AESCipher()
 
-    with open(dst_path, "wb") as new_file:
+    with open(file_path, "wb") as new_file:
         key_filename = aes.write_chiper_text(DEFAULT_FILE_CONTENT, new_file, file_name)
 
     logger.info("File {} was created".format(file_name))
@@ -70,16 +67,16 @@ def create_file():
     return file_name, key_filename
 
 
-def get_metadata(name):
+def get_metadata(file_name):
     """Return stat object with metadata inside
 
-    :param name: string with <file name + file extension>
+    :param file_name: string with <file name + file extension>
     :return: stat obejct with metadata
     """
-    file_path = check_file_existence(name)
-    return OrderedDict(
-        name=name,
-        create_date=convert_datetime(os.path.getctime(file_path)),
-        size=os.path.getsize(file_path),
-        content=read_file(name)
-    )
+    file_path = os.path.join(os.getenv(ENVVAR_NAME_ROOT), file_name + FILE_EXTENSION)
+    check_file_existence(file_path)
+
+    return {"name": file_name,
+            "create_date": convert_datetime(os.path.getctime(file_path)),
+            "size": os.path.getsize(file_path),
+            "content": read_file(file_name)}
